+++ src/realm/sync/subscriptions.cpp
@@ -674,10 +674,20 @@ SubscriptionStore::SubscriptionStore(DBRef db, util::UniqueFunction<void(int64_t
         load_sync_metadata_schema(tr, &internal_tables);
     }
 
-    // There should always be at least one subscription set so that the user can always wait for synchronizationon
-    // on the result of get_latest().
-    if (auto sub_sets = tr->get_table(m_sub_set_table); sub_sets->is_empty()) {
+    // Make sure the subscription set table is properly initialized
+    initialize_subscriptions_table(std::move(tr), false);
+}
+
+void SubscriptionStore::initialize_subscriptions_table(TransactionRef&& tr, bool clear_table)
+{
+    if (auto sub_sets = tr->get_table(m_sub_set_table); clear_table || sub_sets->is_empty()) {
         tr->promote_to_write();
+        // If erase_table is true, clear out the sub_sets table
+        if (clear_table) {
+            sub_sets->clear();
+        }
+        // There should always be at least one subscription set so that the user can always wait
+        // for synchronizationon on the result of get_latest().
         auto zero_sub = sub_sets->create_object_with_primary_key(Mixed{int64_t(0)});
         zero_sub.set(m_sub_set_state, static_cast<int64_t>(SubscriptionSet::State::Pending));
         zero_sub.set(m_sub_set_snapshot_version, tr->get_version());
@@ -796,6 +806,42 @@ std::vector<SubscriptionSet> SubscriptionStore::get_pending_subscriptions() cons
         subscriptions_to_recover.push_back(get_by_version(cur_query_version));
     }
     return subscriptions_to_recover;
+}
+
+void SubscriptionStore::notify_all_state_change_notifications(Status status)
+{
+    std::unique_lock<std::mutex> lk(m_pending_notifications_mutex);
+    m_pending_notifications_cv.wait(lk, [&] {
+        return m_outstanding_requests == 0;
+    });
+
+    auto to_finish = std::move(m_pending_notifications);
+    lk.unlock();
+
+    // Just complete/cancel the pending notifications - this function does not alter the
+    // state of any pending subscriptions
+    for (auto& req : to_finish) {
+        req.promise.set_error(status);
+    }
+}
+
+void SubscriptionStore::terminate()
+{
+    // Clear out and initialize the subscription store
+    initialize_subscriptions_table(m_db->start_read(), true);
+
+    std::unique_lock<std::mutex> lk(m_pending_notifications_mutex);
+    m_pending_notifications_cv.wait(lk, [&] {
+        return m_outstanding_requests == 0;
+    });
+    auto to_finish = std::move(m_pending_notifications);
+    m_min_outstanding_version = 0;
+
+    lk.unlock();
+
+    for (auto& req : to_finish) {
+        req.promise.emplace_value(SubscriptionSet::State::Superseded);
+    }
 }
 
 MutableSubscriptionSet SubscriptionStore::get_mutable_by_version(int64_t version_id)